requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite
pydantic==2.4.2
asyncpg==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.6
numpy
psycopg2-binary
httpx
pytest
pytest-asyncio
openai
